Pretty css.
Hi, you must be wondering right about now.. What the actual fuck is this? Well... It's me, The Depression Goebie. 
My CSS skills aint that good, but I felt your thingy needed some pizzaz! Sooo, I did it. I added it. I even tried to make the filesizes as small as possible. Except for that one font, cba with that one.
I sorta kinda hope you like it? if not, well.. It was time well spent.

^^,/